+++ lib/access_quill_manager.dart
@@ -1,1 +1,3 @@
 export 'src/impl/access_quill_manager.dart';
+export 'src/interface/quill.dart';
+export 'src/interface/quill_manager.dart';