+++ lib/console_quill.dart
@@ -1,2 +1,3 @@
 export 'src/impl/console_quill/color_style/color_style.dart';
 export 'src/impl/console_quill/console_quill.dart';
+export 'src/interface/quill.dart';