+++ lib/file_quill.dart
@@ -1,0 +1,2 @@
+export 'src/impl/file_quill.dart';
+export 'src/interface/quill.dart';