+++ lib/quick_quotes_quill.dart
@@ -1,4 +1,4 @@
 library quick_quotes_quill;
 
-export 'src/interface/quill_base.dart';
-export 'src/interface/quill_manager_base.dart';
+export 'src/interface/quill.dart';
+export 'src/interface/quill_manager.dart';