+++ lib/src/impl/console_quill/quill_tag.dart
@@ -1,5 +1,0 @@
-enum QuillTag {
-  log,
-  error,
-  info,
-}